Option to delete missing mails from backup folder
Emails which have been backed-up are retained in the backup even if they
are deleted from the account at a later date.  This may be undesirable
as the deleted mails will re-appear when restoring from the backup.

To solve this issue, add an option (disabled by default) to delete
previously backed-up emails if they are no-longer returned by the backup
search query.