Skip live API tests and Drive upload for forks
The 'Live API tests push only' and 'Upload to Google Drive, build only'
steps of the GitHub Actions workflow fail on forks as the PASSCODE and
GHCLIENT secrets are unavailable.  Add tests for the value of
repository_owner on these steps to allow the workflow to complete
successfully for forks.